gui_gtk_x11: make bypassing the ASCII glyph cache user-settable
This patch introduces a variable, g:gtk_nocache which (if set) contains
four integers forming a bit map that specifies for which characters <
128 the ASCII glyph cache should be bypassed, and proper glyph shaping
procedures applied instead. If the variable is unset, the glyph cache is
used unconditionally. If set to e.g. [0x00000000, 0xfc00ffff,
0xf8000001, 0x78000001], this will bypass the glyph cache for everything
but control characters (< 32), [0-9A-Za-z], and ASCII code 127. When
using the GUI with a font like Hasklig or PragmataPro, this will enable
most of the programming ligatures (rendering artefacts from
redrawing/cursor movement will still occur, ^L fixes those).